Handle dicts as lists of tuples
Previous commit changed the way sets were handled. Since sets cannot be
used as dict values, they have been replaced as lists internally.

There is, however, the opposite side of the issue: sets or arrays being
used as *keys* in the dict. It is absolutely valid RESP3-wise, but
invalid from Python perspective as sets and lists are not hashable.

This commit changes the way dicts are handled by making them lists of
tuples instead. The library client is supposed to handle the conversion
into proper types themselves.

As previous commit, this commit introduces a breaking change.

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>