reader: check result of Python API in tryParentize
If the code attempts to insert a map as a set item or any unhashable
object as a dict key, `PyDict_SetItem` or `PySet_Add` return `-1`
and set the exception correspondinly.
Ignoring this return value can lead to extension crash.

This commit adds checks of return value of both functions so that a
call to `Reader.gets()` would fail with an exception instead of
crashing.

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>

Check result of PySet_Add as well