reader: add new keywordarg `listOnly`
The new argument changes the way array objects are produced. If
`listOnly` is set to `True`, sets will be returned as Python lists, and
maps will be returned as lists of tuples instead of Python dicts.

This mode should be helpful for cases when RESP3 message contains data
that is not valid from Python's perspective, like a map as a set item or
list/array/set as a map's key.

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>