MANIFEST: add exported headers directory
When hiredis was forked to libvalkey, it changed its layout to have a
proper `src` and `include` directories. Upon forking hiredis-py to
libvalkey-py, `include` directory was missing from the source package
which means the module could not be built from the sdist package.
This issue was hidden by the fact that all the supported systems were
provided with a binary package hence noone actually needed to build it.

This was revealed when people started using libvalkey-py with Python
3.13 that was missing a binary package.

This commit fixes this bug by adding `include` directory to MANIFEST.in

Fixes #19

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>